chore(build): move scripts
move scripts to subfolder

Closes #28
