+++ package.json
@@ -9,15 +9,15 @@
     "build": "npm run build:webpack",
     "prebuild:esm": "npm run prevendor && npm run prescript && npm run preserviceWorker && npm run prescss && npm run prefavicon",
     "build:esm": "npm run vendor && npm run script && npm run scss && npm run favicon && npm run serviceWorker",
-    "postbuild:esm": "ts-node minify.ts",
+    "postbuild:esm": "ts-node ./scripts/minify.ts",
     "prebuild:webpack": "rimraf ./dist",
     "build:webpack": "cross-env NODE_ENV=production webpack --config webpack.config.js",
     "test": "jest --runInBand --env=enzyme --setupFilesAfterEnv=jest-enzyme --reporters=jest-spec-reporter jest-junit --coverage --coverageDirectory ../coverage --rootDir ./src",
     "test:update": "npm run test -- -u",
     "e2e:serve": "webpack-dev-server --config ./e2e/webpack.config.js --quiet true --liveReload false --inline false",
-    "e2e": "cross-env-shell NODE_PORT=$(ts-node findport.ts) NODE_E2E=webpack jest --testTimeout=60000 --runInBand --reporters=jest-spec-reporter --preset=jest-puppeteer --coverage --coverageDirectory ../coverage-e2e --rootDir ./e2e",
-    "serve": "ts-node server.ts",
-    "e2e:esm": "cross-env-shell NODE_PORT=$(ts-node findport.ts) NODE_E2E=esm jest --testTimeout=60000 --runInBand --reporters=jest-spec-reporter --preset=jest-puppeteer --coverage --coverageDirectory ../coverage-e2e --rootDir ./e2e",
+    "e2e": "cross-env-shell NODE_PORT=$(ts-node ./scripts/findport.ts) NODE_E2E=webpack jest --testTimeout=60000 --runInBand --reporters=jest-spec-reporter --preset=jest-puppeteer --coverage --coverageDirectory ../coverage-e2e --rootDir ./e2e",
+    "serve": "ts-node ./scripts/server.ts",
+    "e2e:esm": "cross-env-shell NODE_PORT=$(ts-node ./scripts/findport.ts) NODE_E2E=esm jest --testTimeout=60000 --runInBand --reporters=jest-spec-reporter --preset=jest-puppeteer --coverage --coverageDirectory ../coverage-e2e --rootDir ./e2e",
     "livereload": "livereloadx -a ./public",
     "prescript": "rimraf ./public/script",
     "script": "tsc --project ./src/script --outDir ./public/script",
@@ -28,7 +28,7 @@
     "prewatch-scss": "npm run scss",
     "watch-scss": "node-sass --output-style compressed  --watch --output ./public ./src/style/styles.scss",
     "prefavicon": "rimraf ./public/assets ./public/favicon.html",
-    "favicon": "ts-node favicon.ts",
+    "favicon": "ts-node ./scripts/favicon.ts",
     "prevendor": "rimraf ./public/vendor ./public/es-module-shims.js",
     "vendor": "rimraf ./public/vendor && npm run vendor:base && npm run vendor:faker && npm run vendor:fontawesome && npm run vendor:react && npm run vendor:recharts && npm run vendor:esmoduleshims",
     "vendor:base": "cross-env NODE_ENV=production rollup -c ./src/vendor/rollup-base.config.js",
