+++ scripts/findport.ts
@@ -1,4 +1,4 @@
 import { getPort } from 'portfinder-sync';
-import { baseport } from './package.json';
+import { baseport } from '../package.json';
 
 console.log(getPort(baseport));
