+++ scripts/minify.ts
